requirements.txt
# Azure Agentic RAG System Requirements (Python 3.12 Compatible)

# Core Web Framework
streamlit>=1.30.0

# Azure Services
azure-search-documents>=11.4.0
azure-core>=1.30.0
azure-identity>=1.15.0

# OpenAI and AI Services
openai>=1.12.0

# LangChain Framework (Latest versions for Python 3.12)
langchain>=0.1.0
langchain-openai>=0.0.8
langgraph>=0.0.30
langchain-community>=0.0.20

# Essential Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0

# Data Processing (Python 3.12 Compatible)
numpy>=1.26.0
pandas>=2.2.0

# Async and HTTP
aiohttp>=3.9.0
httpx>=0.27.0

# Logging
loguru>=0.7.2

# Development and Testing
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0

# Type Support
typing-extensions>=4.9.0

# Additional Production Support
uvicorn>=0.27.0
psutil>=5.9.0